+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 #3.7.0
 - New Features
   - @AllowUnusedBindings SqlObject customizer, like SqlStatements.setUnusedBindingsAllowed
+  - Enums config class to change the default policy for binding and mapping Enum values.
+  - @UseEnumStrategy SqlObject configurer for setting the default enum strategy.
+  - @EnumByName and @EnumByOrdinal qualifying annotations to override said default policy.
 
 #3.6.0
 - New Features